Use smallest numeric literal at the sql parser
This change will use the smallest numeric literal when converting untyped literal
inputs at the sql parser. Till now a numeric literal input resulted either in a
`LongLiteral` or `DoubleLiteral`.
A special logic inside the (old) function parameter binding tried to downcast
these literals later on which resulted in bugs on nested function expressions.
See https://github.com/crate/crate/issues/9652.
Using smallest possible numeric literal at the parser will supersede such logic
while ensuring correct function matching also on nested expressions.

+++ server/src/test/java/io/crate/analyze/GroupByAnalyzerTest.java
@@ -168,7 +168,7 @@ public class GroupByAnalyzerTest extends CrateDummyClusterServiceUnitTest {
         QueriedSelectRelation relation = analyze("select 58 as fiftyEight from foo.users group by fiftyEight;");
         assertThat(relation.groupBy().isEmpty(), is(false));
         List<Symbol> groupBySymbols = relation.groupBy();
-        assertThat(groupBySymbols.get(0), isAlias("fiftyeight", isLiteral(58L)));
+        assertThat(groupBySymbols.get(0), isAlias("fiftyeight", isLiteral((short) 58)));
     }
 
     @Test
